pyproject.toml
[project]
name = "basemesh"
version = "0.1.0"
description = "Send Base (Ethereum L2) transactions over Meshtastic/LoRa mesh networks"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "meshtastic>=2.0",
    "web3>=6.0",
    "eth-account>=0.11.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "cryptography>=41.0",
    "mnemonic>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
basemesh = "basemesh.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/basemesh"]
